pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nichescope"
version = "0.1.0"
description = "Creator Intelligence Platform — content gaps, competitor radar, and performance insights for YouTubers"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    "python-telegram-bot>=20.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "feedparser>=6.0",
    "scikit-learn>=1.3",
    "apscheduler>=3.10",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.25",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.setuptools]
packages = ["nichescope"]

[tool.setuptools.package-data]
nichescope = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
